type_mat3x3: work around major Intel CPU bug
This is on the Intel Core Ultra 7 155H.

For whatever reason, the existing implementation of mat3*vec3 ended up
being MASSIVELY slower than a mat4*vec4:

    mat4*vec4 : 1050.19 M/sec
    mat3*vec3 : 9.8 M/sec

That is not a typo. It is consistent across GCC and Clang, for either
Clang vectors or for SSE4.2 intrinsics.

Switching to a row-based approach greatly improves the performance on
Intel.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>